Optimized the package files (#9)
To optimize the package, a files field was added to the package.json,
specifying only essential files for publication: README.md,
CHANGELOG.md, incremental-build.mjs, the bin directory, and
package.json. This ensures that only the necessary files are included,
reducing the package size and making it cleaner and more efficient for
end-users. Verification was done using npm pack --dry-run to confirm the
correct files are packaged.

Fixes: #8